pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compiler"
version = "0.0.1"
dependencies = [
	"antlr4-python3-runtime==4.13.1",
]

[project.scripts]
codegen = "compiler.CodeGen:main"
interpreter = "compiler.Interpreter:main"
